rbd-mirror: renamed 'types' files to 'Types'
