+++ src/tools/rbd_mirror/CMakeLists.txt
@@ -20,7 +20,7 @@ set(rbd_mirror_internal
   PoolWatcher.cc
   ServiceDaemon.cc
   Threads.cc
-  types.cc
+  Types.cc
   image_deleter/RemoveRequest.cc
   image_deleter/SnapshotPurgeRequest.cc
   image_deleter/TrashMoveRequest.cc
