+++ src/tools/rbd_mirror/Types.cc
@@ -1,7 +1,7 @@
 // -*- mode:C++; tab-width:8; c-basic-offset:2; indent-tabs-mode:t -*-
 // vim: ts=8 sw=2 smarttab
 
-#include "types.h"
+#include "tools/rbd_mirror/Types.h"
 
 namespace rbd {
 namespace mirror {
